requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain>=0.0.300
langchain-community
langchain-core
langchain-chroma
langchain-groq
langchain_huggingface
chromadb>=0.3.27
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=1.13.0
pymupdf
python-dotenv
requests
tqdm
pydantic
